# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "SplitWiseMoney API",
                 "description":  "SplitWiseMoney - Expense sharing web application API documentation",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/notifications/mark-read":  {
                                                       "post":  {
                                                                    "tags":  [
                                                                                 "Notifications"
                                                                             ],
                                                                    "summary":  "Mark all notifications for the current user as read",
                                                                    "operationId":  "markAllAsRead",
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "OK",
                                                                                                  "content":  {
                                                                                                                  "*/*":  {
                                                                                                                              "schema":  {
                                                                                                                                             "type":  "string"
                                                                                                                                         }
                                                                                                                          }
                                                                                                              }
                                                                                              }
                                                                                  },
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ]
                                                                }
                                                   },
...
                  "/api/notifications":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "Notifications"
                                                                  ],
                                                         "summary":  "Get paginated user notifications",
                                                         "operationId":  "getUserNotifications",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "page",
                                                                                "in":  "query",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int32",
                                                                                               "default":  0
                                                                                           }
                                                                            },
                                                                            {
                                                                                "name":  "size",
                                                                                "in":  "query",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int32",
                                                                                               "default":  20
                                                                                           }
                                                                            },
                                                                            {
                                                                                "name":  "sort",
                                                                                "in":  "query",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "default":  "createdAt,desc"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/PageNotificationDto"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   }
                                                                       },
                                                         "security":  [
                                                                          {
                                                                              "bearerAuth":  [

                                                                                             ]
                                                                          }
                                                                      ]
                                                     }
                                         },
                  "/api/notifications/unread-count":  {
                                                          "get":  {
                                                                      "tags":  [
                                                                                   "Notifications"
                                                                               ],
                                                                      "summary":  "Get count of unread notifications for current user",
                                                                      "operationId":  "getUnreadCount",
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK",
                                                                                                    "content":  {
                                                                                                                    "*/*":  {
                                                                                                                                "schema":  {
                                                                                                                                               "type":  "integer",
                                                                                                                                               "format":  "int64"
                                                                                                                                           }
                                                                                                                            }
                                                                                                                }
                                                                                                }
                                                                                    },
                                                                      "security":  [
                                                                                       {
                                                                                           "bearerAuth":  [

                                                                                                          ]
                                                                                       }
                                                                                   ]
                                                                  }
                                                      },
                  "/api/history":  {
                                       "get":  {
                                                   "tags":  [
                                                                "History"
                                                            ],
                                                   "summary":  "Get paginated user activity logs history",
                                                   "operationId":  "getUserLogs",
                                                   "parameters":  [
                                                                      {
                                                                          "name":  "page",
                                                                          "in":  "query",
                                                                          "required":  false,
                                                                          "schema":  {
                                                                                         "type":  "integer",
                                                                                         "format":  "int32",
                                                                                         "default":  0
                                                                                     }
                                                                      },
                                                                      {
                                                                          "name":  "size",
                                                                          "in":  "query",
                                                                          "required":  false,
                                                                          "schema":  {
                                                                                         "type":  "integer",
                                                                                         "format":  "int32",
                                                                                         "default":  20
                                                                                     }
                                                                      },
                                                                      {
                                                                          "name":  "sort",
                                                                          "in":  "query",
                                                                          "required":  false,
                                                                          "schema":  {
                                                                                         "type":  "string",
                                                                                         "default":  "createdAt,desc"
                                                                                     }
                                                                      }
                                                                  ],
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK",
                                                                                 "content":  {
                                                                                                 "*/*":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/PageActivityLogDto"
                                                                                                                        }
                                                                                                         }
                                                                                             }
                                                                             }
                                                                 },
                                                   "security":  [
                                                                    {
                                                                        "bearerAuth":  [

                                                                                       ]
                                                                    }
                                                                ]
                                               }
                                   },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "NotificationDto":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "id":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int64"
                                                                                         },
                                                                                  "message":  {
                                                                                                  "type":  "string"
                                                                                              },
                                                                                  "type":  {
                                                                                               "type":  "string"
                                                                                           },
                                                                                  "isRead":  {
                                                                                                 "type":  "boolean"
                                                                                             },
                                                                                  "createdAt":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "date-time"
                                                                                                }
                                                                              }
                                                           },
                                       "PageNotificationDto":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "totalPages":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int32"
                                                                                                     },
                                                                                      "totalElements":  {
                                                                                                            "type":  "integer",
                                                                                                            "format":  "int64"
                                                                                                        },
                                                                                      "size":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int32"
                                                                                               },
                                                                                      "content":  {
                                                                                                      "type":  "array",
                                                                                                      "items":  {
                                                                                                                    "$ref":  "#/components/schemas/NotificationDto"
                                                                                                                }
                                                                                                  },
                                                                                      "number":  {
                                                                                                     "type":  "integer",
                                                                                                     "format":  "int32"
                                                                                                 },
                                                                                      "first":  {
                                                                                                    "type":  "boolean"
                                                                                                },
                                                                                      "last":  {
                                                                                                   "type":  "boolean"
                                                                                               },
                                                                                      "numberOfElements":  {
                                                                                                               "type":  "integer",
                                                                                                               "format":  "int32"
                                                                                                           },
                                                                                      "sort":  {
                                                                                                   "$ref":  "#/components/schemas/SortObject"
                                                                                               },
                                                                                      "pageable":  {
                                                                                                       "$ref":  "#/components/schemas/PageableObject"
                                                                                                   },
                                                                                      "empty":  {
                                                                                                    "type":  "boolean"
                                                                                                }
                                                                                  }
                                                               },
...
                                       "PageActivityLogDto":  {
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "totalPages":  {
                                                                                                        "type":  "integer",
                                                                                                        "format":  "int32"
                                                                                                    },
                                                                                     "totalElements":  {
                                                                                                           "type":  "integer",
                                                                                                           "format":  "int64"
                                                                                                       },
                                                                                     "size":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int32"
                                                                                              },
                                                                                     "content":  {
                                                                                                     "type":  "array",
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ActivityLogDto"
                                                                                                               }
                                                                                                 },
                                                                                     "number":  {
                                                                                                    "type":  "integer",
                                                                                                    "format":  "int32"
                                                                                                },
                                                                                     "first":  {
                                                                                                   "type":  "boolean"
                                                                                               },
                                                                                     "last":  {
                                                                                                  "type":  "boolean"
                                                                                              },
                                                                                     "numberOfElements":  {
                                                                                                              "type":  "integer",
                                                                                                              "format":  "int32"
                                                                                                          },
                                                                                     "sort":  {
                                                                                                  "$ref":  "#/components/schemas/SortObject"
                                                                                              },
                                                                                     "pageable":  {
                                                                                                      "$ref":  "#/components/schemas/PageableObject"
                                                                                                  },
                                                                                     "empty":  {
                                                                                                   "type":  "boolean"
                                                                                               }
                                                                                 }
                                                              },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "name":  "bearerAuth",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}